Fix asset loader `getPublicURL()` public path formatting issue
This PR fixes an issue regarding the asset loader getPublicURL() method not
correctly formatting the url when the public path has not been specified. In
such a situation, it prefixes the path with a forward slash. Actual usage,
however, has revealed that rather than prefixing the url with a forward slash,
it should instead return the asset name as is, when the public path has not been set.