Add support for vanity URL header to navigation middleware
The navigation data usually references vanity URLs configured using the URL management service so this
change enables us to set the vanity URL as the current path if it is available.

Please note that the vanity URL is exposed as a header because the CDN (or router when running an app locally)
will rewrite a request to use the underlying URL if a vanity is configured for it. This means our apps only
need to be configured to serve the original URL and aren't usually aware if a vanity is used.