Use inline Babel helpers to prevent bundle hashes being update by app-level code changes.
You might think we'd want to abstract the helper functions so they can be reused but doing so
means we generate unstable hashes because the generated helper modules are at the bottom of
the dependency tree but their contents depends on the features each app uses. Inlining them
adds little (usually <1kb) to our total JS payload because the minimizer can usually reduce
them down and actually result in fewer function calls overall!