Replace usage of 'shared ui' with 'global ui'
Words are hard! I'm trying to clarify the scope of the UI components which should be included
within Page Kit. Currently we use "shared" but this may imply a component used by a few apps,
but the rule we've been working to is to include only what is needed for _most_ apps.

The header, navigation, footer, and "core branding" are _global_ and intrinsic to our brand so
I have updated the wording across the readmes to reflect this.