Remove incremental compiler option from root TS config to prevent conflict with --noEmit flag
Since TS v3.8.2 the `--noEmit` flag cannot be used in conjuction with the `incremental` compiler option.

https://github.com/microsoft/TypeScript/issues/32882

We currently use the `--noEmit` flag to quickly run a type check on pre-commit and the `incremental` option for all our TS to JS compilation.

I'd like to maintain this setup so that we do not have to compile code on commit, nor slow down builds by forcing them to run from scratch each time.

To do this I have removed the incremental flag from the root TS config and appended the flag to each `tsc` script specified by each package.