Refactor flags to align with appContext logic
Since `dotcom-ui-app-context` and `dotcom-ui-flags` share functionality this commit moves JSON output into the embed component which is the same as how it works in app-context. This'll hopefully make it easier to use the new package in the future.

May be a breaking change if formatFlagsJSON is part of the public API.