Add support for right aligned sub navigation in `dotcom-ui-header`
Adds a `subsectionsRight` array to the `THeaderProps` along with the
existing `subsections` array, with items which are positioned to the
right of the subnav. This will be used my `next-profile` initially
to align a signout link to the right of the subnav. There is no
requirement for this to be defined in navigation apis (Origami
Navigation Service or Next Navigation API) but could be added
in the future without a breaking change to the APIs.

Related o-header PR: https://github.com/Financial-Times/o-header/pull/395
Related issue: https://github.com/Financial-Times/o-header/issues/376