Pin sass dependency
[sass version 1.32.0](https://github.com/sass/dart-sass/releases/tag/1.32.0)
introduced a deprecation warning for passing non-% numbers as lightness
and saturation to `hsl()`. This function is used by `o-colors`.

When building `next-article`, this warning is outputted in the console
over _many pages_ and it flickers really badly for a while. For people
sensitive to page flickering, this is not great. To prevent this, we're
thinking about pinning the dependency to before the message was
introduced. Thoughts?

You can replicate this by running the following commands in
`next-article`:

```
make clean
make install .env
make build
```