Require the latest o-typography/o-fonts release
`dotcom-page-kit` preloads fonts from `o-fonts-assets` however
the version to use is hard-coded and not tied to the actual
version of `o-fonts` being installed. This update makes sure the
fonts which are preloaded and used are aligned, as a new version
of `o-fonts-assets` has been released.

Ideally we'd add a JS API to Origami, in addition to the o-font
Sass variable, so `dotcom-page-kit` can grab the correct font
urls to preload. This might tie in to some future work the
core team are considering to define variables in a JS/JSON format
which compile to Sass (a.k.a design tokens) so for now let's keep
this in sync manually.

Similar past PR:
https://github.com/Financial-Times/dotcom-page-kit/pull/826