Don't require now unused polyfill runtime in test
The regenerator runtime is used to support polyfills of generator
functions. It stopped being output by Webpack since the last commit to
main, possibly because all dependencies that were using it have been
updated to no longer require it and it is now tree shaken? Fortunately,
after updating our browserslist config, we no longer need the Babel
transform as all of our supported browsers now support generators, so we
can safely assume it is not a required webpack bundle.