build: use version of express types compatible with tsc 3
The latest version of @types/express-serve-static-core causes our
TypeScript builds to fail as it uses syntax that our older version of
tsc can't parse. Let's force it to use an older version to avoid this.

In theory, it would be better to use npm overrides to achieve this, but
I can't get it to work in practice as they always seem to fail to handle
one instance of a transitive dependency or other.

We could also work around this by updating to the latest version of tsc
– which might be worthwhile but could introduce some other bugs I don't
want to search for right now, or finally add a package-lock.json file to
this monorepo. I think npm workspaces would be a suitable drop-in for
the athloi workflow we have at the moment at this point.