Ensure that all maps are called on an array
This isn't necessarily the correct solution, but wanted to get a PR open
so we can discuss as a group how to solve this. I've gone with something
very basic and we may want to rethink this module more.

There are cases where the Next Navigation Service will respond with
`null` instead of an array. In these cases I'd prefer if we skip that
part of the navigation rather than the page erroring completely which is
what we saw in the incident earlier today.

Please tear this down if there are better solutions, but I think we need
to start treating data from services that even we maintain as
potentially malformed - Types alone aren't really enough for this case
and I think we should be more defensive against runtime data issues.