refactor(dotcom-ui-header): stop casting undefined and null to a string
We decided that a nullish coalescing operator would be better in this
case, because it maintains the existing behaviour: if the item URL is
null then the attribute is not output. The previous code would have
resulted in the href attributes being set to the string "null".

See https://github.com/Financial-Times/dotcom-page-kit/pull/956#discussion_r995479248

Co-Authored-By: Alex Muller <alexmuller@users.noreply.github.com>
Co-Authored-By: Andy Gout <andygout@users.noreply.github.com>
Co-Authored-By: Ivo Murrell <ivomurrell@users.noreply.github.com>
Co-Authored-By: Joel Carr <joelcarr@users.noreply.github.com>