chore: pin type definition version to support ts 3.x
Seeing as we still aren't using a lockfile for this repository we're
occasionally bumping into @types definitions which are doing minor
version updates that are breaking TypeScript 3.x support. We need to pin
the version of @types/babel__traverse that Jest uses in order to avoid
one such compatibility issue.