refactor: use npm workspaces to publish to npm
Athloi would be running the npm version and npm publish script so we are
now doing that directly rather than through Athloi. We are removing the
use the get-npm-dist-tag script as we do not use anything similar in our
other npm packages (components) like dotcom-tool-kit. To do this, we add
a new filter for prerelease versions which will tag the release as a
prerelease and then duplicate the publish workflow making the amends
needed.