fix: publish versions of each package Page Kit dependency
When publishing to the npm registry, we set the version in each Page Kit package dependency. Athloi used to then update each Page Kit package Page Kit dependency to the latest version number. This is missing in the migration to npm workspaces (CPP-1249).

We can use athloi just for the version step. We should then replace this
with release-please.

https://financialtimes.atlassian.net/browse/CPP-1330