fix: correctly point to declaration files instead of TypeScript source
The types field in package.json is meant to reference the declaration
files that are generated by TypeScript and then can be consumed by it
when installing published packages. Instead of pointing to the
declaration files, we were instead pointing to the original TypeScript
source code which we are (inappropriately) also publishing. This has the
effect that the source code will be transpiled again by TypeScript apps
that depend on the Page Kit packages.