build: make sure server packages are built before dependent middleware
Now that we're correctly pointing to the generated declaration files
rather than the original TypeScript source, we have to make sure the
declaration files have actually been generated before we perform type
checking on packages that depend on them. We can enforce an order to
which packages are built by using references to specify when a package
depends on another in the monorepo.

I've only added the references that were causing an error when running a
clean build for now. We might run non-deterministically run into other
build errors when making further changes to the code in the future.