feat: Replace the settings & account link with myft - AB test
As part of an AB test we want the ability to change the current "settings & account" link to be a myFT link instead

As this data comes from an external service we have achieved this in a hacky way for now. If the test is successful then we would look to update the data.