fix: Remove unnecessary peer dependencies (n-topic-search) for code splitting
This peer dependencies were added in https://github.com/Financial-Times/dotcom-page-kit/pull/1012

However I don't believe they are needed as peer dependencies at all as the code splitting package doesn't use them.

It only references them so if they are used in a consuming app the code can be split into their own bundle as they are commonly used.

The issue with including these as a peer dependencies is that it forces apps to include these dependencies at a set version regardless.