Update header nav right to use new data structure
Previously there was 2 objects

1. navbar-right
2. navbar-right-anon

These sound like they are data for the same area of the header just based on different user states (anon vs logged in)

However they were being used in different placements within the header.

navbar-right = This is used in the second/bottom row of the header
navr-bar-right-non = This is used in the first/top row of the header

This causes confusion when trying to make changes to the navigation data and header.

To get around this we are adding a 2 new objects

1. navbar-top-right
2. navbar-top-right-anon

These should always in the top/first row of the header.