feat: remove postcss-import from Sass plugin
With Webpack 5 we've been able to update to a newer version of the
`sass-loader` loader. The newer version no longer requires imports to be
prefixed with a `~` if they want to use Webpack's resolution logic
(i.e., traversing `node_modules`.) This is particularly good for
stylesheets that import stylesheets that themselves import stylesheets,
as the transitive imports are no longer required to use the non-standard
`~` prefix to work.

This means we no longer need to use the `postcss-import` plugin to work
around the import behaviour. Importing using `sass-loader` makes native
use of Webpack's resolution logic so should be better equipped to handle
de-duping etc., though that has not been thoroughly tested: the main
motivation here is to use fewer plugins and be more native.