build: include missing optional sass-embedded packages in lockfile
For some reason a previous npm installation didn't include all the
optional dependencies for the OS specific versions of the sass-embedded
executable – only the macOS version which would have been installed on
the host device. The other resolutions are needed if we want to be able
to install on other operating systems, such as our CI server.

I deleted the lockfile and reinstalled to get the resolutions required.
I then copied them back into the previous lockfile to minimise package
updates. It does mean we're using a newer minor version of sass-embedded
(and its OS executables) now but that's fine.