feat: add ownership attributes to header and footer
This will allow users of Reliability's ownership inspector to see which
parts of the header and footer are owned by which packages.

This is the first step towards creating a real-world example of
ownership attributes. These will be included in next-static pages which
can be used to demonstrate the ownership inspector.